refactor(docs): accordion codes
