+++ apps/docs/content/components/accordion/bordered-variant.raw.jsx
