+++ apps/docs/content/components/accordion/compact.raw.jsx
