+++ apps/docs/content/components/accordion/compact.ts
@@ -1,4 +1,4 @@
-import App from "./compact.raw.tsx?raw";
+import App from "./compact.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
